+++ echelon/notices.php
@@ -60,8 +60,12 @@ require 'app/views/global/header.php';
 if(!$db->error) :
 ?>
 
+<div class="page-header">
+	<h1>Notices</h1>
+	<p>There are a total of <span class="badge"><?php echo $total_rows; ?></span> notices, made by admins in the server(s)</p>
+</div>
+
 <table class="table table-striped table-hover" summary="A list of <?php echo limit_rows; ?> notices made by admins in the server regarding a certain player">
-	<caption>Notices<small>There are a total of <strong><?php echo $total_rows; ?></strong> notices, made by admins in the server(s)</small></caption>
 	<thead>
 		<tr>
 			<th>Name</th>